requirement.txt
Flask==2.2.5
click==8.1.7
psycopg2-binary==2.9.9

